invoicing/requirements.txt
fuzzywuzzy==0.18.0
python-dotenv==1.0.1
reportlab==4.1.0
Requests==2.32.3
WooCommerce==3.0.0
